#165 Apply comments in PR
 <organisation, naming
