+++ src/main/java/nl/tudelft/ti2206/group9/gui/skin/NoobSkin.java
@@ -6,7 +6,7 @@ package nl.tudelft.ti2206.group9.gui.skin;
 public class NoobSkin extends Skin {
 
     /**
-     * Constructor for a Noob currentSkin.
+     * Constructor for a Noob skin.
      */
     public NoobSkin() {
         super(0, "Noob", "noob");
